requirements.txt
# Основные зависимости
python-telegram-bot==20.3
flask==2.2.5
flask-sqlalchemy==2.5.1
langchain==0.0.365
chromadb==0.3.23
transformers==4.34.0
bitsandbytes==0.40.0
torch==2.0.1
sentence-transformers==2.2.2
python-dotenv==1.0.0

# Дополнительные зависимости для обработки документов
unstructured==0.10.30  # Для загрузки PDF/DOCX
pypdf==3.17.0  # Обработка PDF
python-docx==1.1.0  # Обработка DOCX

# Для асинхронной работы и PostgreSQL
psycopg2-binary==2.9.6
asyncpg==0.28.0
asyncio==3.4.3

# Утилиты
python-dateutil==2.8.2
tqdm==4.65.0  # Прогресс-бары